options moved from /pkg to /internal
