+++ cmd/vt/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"github.com/happyhackingspace/vulnerable-target/internal/cli"
 	"github.com/happyhackingspace/vulnerable-target/internal/logger"
-	"github.com/happyhackingspace/vulnerable-target/pkg/options"
+	"github.com/happyhackingspace/vulnerable-target/internal/options"
 	"github.com/happyhackingspace/vulnerable-target/pkg/templates"
 )
